linking critters and endangered redlist status
first use fish_eda notebook to clean and process insects and fish names from nookipedia website to fetch the scientific name using beautiful soup. Then use the critters_endangered notebook to take the results from the fish_eda and link the scientific name on the file to the ICUN red list and fetch the Endangered Category if endangered.